Moved weights calculation function to new module which will contain the full blending plugin combo.
