+++ lib/improver/blending/weights.py
@@ -38,88 +38,7 @@ from scipy.interpolate import interp1d
 
 import iris
 
-from improver.utilities.cube_manipulation import (
-    check_cube_coordinates, sort_coord_in_cube)
-
-
-def calculate_blending_weights(cube, blend_coord, method,
-                               blend_coord_unit=None,
-                               weighting_coord=None, wts_dict=None,
-                               y0val=None, ynval=None, cval=None):
-    """
-    Wrapper for plugins to calculate blending weights using the command line
-    options specified.
-
-    Args:
-        cube (iris.cube.Cube):
-            Cube of input data to be blended
-        blend_coord (str):
-            Coordinate over which blending will be performed (eg "model" for
-            grid blending)
-        method (str):
-            Weights calculation method ("linear", "nonlinear", "dict" or
-            "mask")
-
-    Kwargs:
-        blend_coord_unit (str or cf_units.Unit):
-            Unit of blending coordinate (for default weights plugins)
-        weighting_coord (str):
-            Coordinate over which linear weights should be calculated from dict
-        wts_dict (dict):
-            Dictionary containing parameters for linear weights calculation
-        y0val (float):
-            Intercept parameter for default linear weights plugin
-        ynval (float):
-            Gradient parameter for default linear weights plugin
-        cval (float):
-            Parameter for default non-linear weights plugin
-
-    Returns:
-        weights (iris.cube.Cube):
-            Cube containing 1D array of weights for blending
-    """
-    # sort input cube by blending coordinate
-    cube = sort_coord_in_cube(cube, blend_coord, order="ascending")
-
-    # set blending coordinate units
-    if "time" in blend_coord:
-        coord_unit = cf_units.Unit(blend_coord_unit, "gregorian")
-    elif blend_coord_unit != 'hours since 1970-01-01 00:00:00.':
-        coord_unit = blend_coord_unit
-    else:
-        coord_unit = 'no_unit'
-
-    # calculate blending weights
-    if method == "dict":
-        # get dictionary access
-        if "model" in blend_coord:
-            config_coord = "model_configuration"
-        else:
-            config_coord = blend_coord
-
-        # calculate linear weights from dictionary
-        weights_cube = ChooseWeightsLinear(
-            weighting_coord, wts_dict,
-            config_coord_name=config_coord).process(cube)
-
-        # sort weights cube by blending coordinate
-        # TODO do we need this now?  Check / match order in blending plugin...
-        weights = sort_coord_in_cube(
-            weights_cube, blend_coord, order="ascending")
-
-    elif method == "linear":
-        weights = ChooseDefaultWeightsLinear(
-            y0val=y0val, ynval=ynval).process(
-                cube, blend_coord, coord_unit=coord_unit)
-
-    elif method == "nonlinear":
-        # this is set here rather than in the CLI arguments in order to check
-        # for invalid argument combinations
-        cvalue = cval if cval else 0.85
-        weights = ChooseDefaultWeightsNonLinear(cvalue).process(
-            cube, blend_coord, coord_unit=coord_unit)
-
-    return weights
+from improver.utilities.cube_manipulation import check_cube_coordinates
 
 
 class WeightsUtilities:
